feat: done input box
