+++ packages/frontend/apps/ios/App/Packages/Intelligents/Sources/Intelligents/Interface/View/InputBox/InputBox.swift
@@ -291,6 +291,13 @@ class InputBox: UIView {
     }
   }
 
+  public func addDocumentAttachment(_ documentAttachment: DocumentAttachment) {
+    performWithAnimation { [self] in
+      viewModel.addDocumentAttachment(documentAttachment)
+      layoutIfNeeded()
+    }
+  }
+
   public var inputBoxData: InputBoxData {
     viewModel.prepareSendData()
   }
