+++ packages/frontend/apps/ios/App/Packages/Intelligents/Sources/Intelligents/Interface/View/MainHeaderView/MainHeaderViewDelegate.swift
@@ -1,0 +1,14 @@
+//
+//  MainHeaderViewDelegate.swift
+//  Intelligents
+//
+//  Created by 秋星桥 on 6/25/25.
+//
+
+import Foundation
+
+protocol MainHeaderViewDelegate: AnyObject {
+  func mainHeaderViewDidTapClose()
+  func mainHeaderViewDidTapDropdown()
+  func mainHeaderViewDidTapMenu()
+}
